+++ hc/api/tests/test_notify_slack.py
@@ -109,3 +109,12 @@ class NotifySlackTestCase(BaseTestCase):
         n = Notification.objects.get()
         self.assertEqual(n.error, "Received status code 404")
         self.assertEqual(mock_post.call_count, 1)
+
+    @patch("hc.api.transports.requests.request")
+    def test_it_disables_channel_on_404(self, mock_post):
+        self._setup_data("123")
+        mock_post.return_value.status_code = 404
+
+        self.channel.notify(self.check)
+        self.channel.refresh_from_db()
+        self.assertTrue(self.channel.disabled)