+++ CHANGELOG.md
@@ -11,6 +11,8 @@ All notable changes to this project will be documented in this file.
 - Add support for Telegram channels (#592)
 - Implement Telegram group to supergroup migration (#132)
 - Update the Slack integration to not retry when Slack returns 404
+- Refactor transport classes to raise exceptions on delivery problems
+- Add Channel.disabled field, for disabling integrations on permanent errors
 
 ### Bug Fixes
 - Fix report templates to not show the "started" status (show UP or DOWN instead)